Ignore `skip_stack_guard_page` on linux for now

+++ lib/engine-jit/src/link.rs
@@ -51,7 +51,7 @@ fn apply_relocation(
         },
         RelocationKind::X86CallPCRel4 => unsafe {
             let (reloc_address, reloc_delta) = r.for_address(body, target_func_address as u64);
-            write_unaligned(reloc_address as *mut usize, reloc_delta as _);
+            write_unaligned(reloc_address as *mut u32, reloc_delta as _);
         },
         RelocationKind::X86PCRelRodata4 => {}
         kind => panic!(
